style: moved forbidden keyword linter to the functioncheck file
