+++ pkg/functioncheck/functioncheck.go
@@ -7,6 +7,7 @@ import (
 	"os/exec"
 	"regexp"
 	"strings"
+	"bufio"
 
 	"github.com/42-Short/shortinette/internal/datastructures"
 	"github.com/42-Short/shortinette/internal/errors"
@@ -35,6 +36,40 @@ func initCompilingEnvironment(allowedItems datastructures.AllowedItems, exercise
 	}
 
 	return nil
+}
+
+func scanForForbiddenKeywords(scanner *bufio.Scanner, forbiddenKeywords []string) (err error) {
+	foundKeywords := make([]string, 0, len(forbiddenKeywords))
+	
+    for scanner.Scan() {
+        word := scanner.Text()
+        for _, keyword := range forbiddenKeywords {
+            if word == keyword {
+                foundKeywords = append(foundKeywords, keyword)
+            }
+        }
+    }
+	if len(foundKeywords) > 0 {
+		return fmt.Errorf("found forbidden keywords: %s", strings.Join(foundKeywords, ", "))
+	}
+    return scanner.Err()
+}
+
+func lintStudentCode(exercisePath string, exerciseConfig datastructures.Exercise) (err error) {
+	file, err := os.Open(exercisePath)
+    if err != nil {
+        return fmt.Errorf("could not open %s: %w", exercisePath, err)
+    }
+    defer file.Close()
+
+    scanner := bufio.NewScanner(file)
+    scanner.Split(bufio.ScanWords)
+
+    err = scanForForbiddenKeywords(scanner, exerciseConfig.ForbiddenKeywords)
+	if err != nil {
+        return err
+    }
+    return nil
 }
 
 func prependHeadersToStudentCode(filePath, exerciseNumber string, exerciseType string, dummyCall string) error {
@@ -143,7 +178,7 @@ func Execute(exerciseConfig datastructures.Exercise) (err error) {
 	}
 
 	exercisePath := fmt.Sprintf("compile-environment/src/%s/%s", exerciseConfig.TurnInDirectory, exerciseConfig.TurnInFile)
-	err = LintStudentCode(exercisePath, exerciseConfig)
+	err = lintStudentCode(exercisePath, exerciseConfig)
 	if err != nil {
 		return err
 	}
